+++ packages/expo-ui/ios/Modifiers/ViewModifierRegistry.swift
@@ -1611,5 +1611,13 @@ extension ViewModifierRegistry {
     register("gridCellAnchor") { params, appContext, _ in
       return try GridCellAnchor(from: params, appContext: appContext)
     }
+
+    register("tag") { params, appContext, _ in
+      return try TagModifier(from: params, appContext: appContext)
+    }
+    
+    register("pickerStyle") { params, appContext, _ in
+      return try PickerStyleModifier(from: params, appContext: appContext)
+    }
   }
 }